+++ graphics/src/types/Vertex1.cpp
@@ -5,11 +5,11 @@
 
 namespace Graphics
 {
-    const uint32_t Vertex1::VERTEX_BYTE_SIZE = sizeof Position + sizeof Color;
+    const size_t Vertex1::VERTEX_BYTE_SIZE = sizeof(Position) + sizeof(Color);
 
-    const uint32_t Vertex1::ELEMENTS_PER_POSITION = sizeof Position / sizeof(float);
-    const uint32_t Vertex1::POSITION_TYPE = GL_FLOAT;
+    const int Vertex1::ELEMENTS_PER_POSITION = sizeof(Position) / sizeof(float);
+    const unsigned int Vertex1::POSITION_TYPE = GL_FLOAT;
 
-    const uint32_t Vertex1::ELEMENTS_PER_COLOR = sizeof Color / sizeof(float);
-    const uint32_t Vertex1::COLOR_TYPE = GL_FLOAT;
+    const int Vertex1::ELEMENTS_PER_COLOR = sizeof(Color) / sizeof(float);
+    const unsigned int Vertex1::COLOR_TYPE = GL_FLOAT;
 }
